# specs/001-media-influence-detection/contracts/api.yaml
openapi: 3.1.0
info:
  title: MITDS API
  description: Media Influence Topology & Detection System REST API
  version: 1.0.0
  contact:
    name: MITDS Team
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.mitds.example/v1
    description: Production
...
paths:
...
  /reports:
    post:
      summary: Generate a report
      tags: [Reports]
      operationId: generateReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
      responses:
        '202':
          description: Report generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncJobResponse'
  /reports/{id}:
    get:
      summary: Get generated report
      tags: [Reports]
      operationId: getReport
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: format
          in: query
          schema:
            type: string
            enum: [json, html, pdf, markdown]
            default: json
      responses:
        '200':
          description: Report content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
            text/html:
              schema:
                type: string
            application/pdf:
              schema:
                type: string
                format: binary

  # ============== INGESTION ==============
...
components:
  schemas:
...
    EvidenceRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        evidence_type:
          type: string
        source_url:
          type: string
          format: uri
        archive_url:
          type: string
          format: uri
        retrieved_at:
          type: string
          format: date-time
...
    ReportRequest:
      type: object
      required: [template_id]
      properties:
        template_id:
          type: string
          enum: [structural_risk, influence_topology, timeline_narrative]
        entity_ids:
          type: array
          items:
            type: string
            format: uuid
        time_range:
          type: object
          properties:
            start:
              type: string
              format: date
            end:
              type: string
              format: date
        options:
          type: object
          additionalProperties: true
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        template_id:
          type: string
        generated_at:
          type: string
          format: date-time
        title:
          type: string
        executive_summary:
          type: string
        methodology:
          type: string
        findings:
          type: array
          items:
            type: object
        limitations:
          type: array
          items:
            type: string
        evidence_refs:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceRef'

    # ============== INGESTION ==============
...
    AsyncJobResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, completed, failed]
        status_url:
          type: string
          format: uri
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []